refactor: Order와 OrderTable 패키지 분리
