+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -9,13 +9,13 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItemQuantity;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.menu.exception.MenuException;
 import kitchenpos.order.exception.OrderException;
-import kitchenpos.order.exception.OrderTableException;
+import kitchenpos.ordertable.exception.OrderTableException;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
