+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,6 +1,7 @@
 package kitchenpos.order.domain;
 
 import kitchenpos.order.exception.OrderException;
+import kitchenpos.ordertable.domain.OrderTable;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
