+++ src/main/java/kitchenpos/ordertable/exception/OrderTableGuestNumberException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.ordertable.exception;
 
 public class OrderTableGuestNumberException extends RuntimeException {
