+++ src/main/java/kitchenpos/ordertable/exception/TableGroupException.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.exception;
+package kitchenpos.ordertable.exception;
 
 public class TableGroupException extends RuntimeException {
