+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -8,16 +8,16 @@ import kitchenpos.order.application.dto.CreateOrderDto;
 import kitchenpos.order.application.dto.CreateOrderLineItemDto;
 import kitchenpos.order.application.dto.OrderDto;
 import kitchenpos.order.application.dto.UpdateOrderStatusDto;
-import kitchenpos.order.domain.GuestNumber;
+import kitchenpos.ordertable.domain.GuestNumber;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItemQuantity;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.order.exception.OrderException;
-import kitchenpos.order.exception.OrderTableException;
+import kitchenpos.ordertable.exception.OrderTableException;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
