+++ src/test/java/kitchenpos/ordertable/application/TableGroupServiceTest.java
@@ -1,19 +1,19 @@
-package kitchenpos.order.application;
+package kitchenpos.ordertable.application;
 
 import kitchenpos.MockServiceTest;
 import kitchenpos.order.application.dto.CreateOrderTableIdDto;
 import kitchenpos.order.application.dto.CreateTableGroupDto;
 import kitchenpos.order.application.dto.OrderTableDto;
 import kitchenpos.order.application.dto.TableGroupDto;
-import kitchenpos.order.domain.GuestNumber;
+import kitchenpos.ordertable.domain.GuestNumber;
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
-import kitchenpos.order.exception.OrderTableException;
-import kitchenpos.order.exception.TableGroupException;
-import kitchenpos.order.application.TableGroupService;
-import kitchenpos.order.domain.OrderTableRepository;
-import kitchenpos.order.domain.TableGroupRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.ordertable.exception.OrderTableException;
+import kitchenpos.ordertable.exception.TableGroupException;
+import kitchenpos.ordertable.application.TableGroupService;
+import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.TableGroupRepository;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
